tests/requirements.txt
selenium-wire
selenium
PyYAML
requests